Added cap puppet:bootstrap task to run bootstrap script (installing ruby, git and puppet) remotely.
This script needs to be run as root; we spent some time trying to figure out how to SSH to all configured servers directly, but capistrano makes this sufficiently odd that we didn't bother. At the moment we are deploying as root; in the future we will move the bootstrap into part of the instance provisioning mechanism, and remove it from here.