The root mysql user should have a password
By adding the password to /root/.my.cnf, any user logged in as root can access mysql tools without specifying the password.  Used an erb template to generate the password, as we could not figure out how to guarantee the availability of a more standard password generation tool before the point where the password is generated.