We could not deploy to machines without the web role.
For the load test, we want the application to be available on a simulator VPS, but this machine should not have the web role. However, only machines with rack installed were provisioned with the directory that the application should be deployed to.

We have more completely extracted the application user and directory from the 'rack' class, into base::application.

We will probably change the deployment directory to something outside of /var/www in the future to reflect this.