Revert this, it doesn't play well with deploy:setup
The capistrano "deploy:setup" task expects to be able to create the main application folder, since it needs to own in to be able to change permissions on it later. Normally this would be owned by the user that logs in for deploy:setup. However, if puppet creates this folder, it will be owned by 'root', breaking capistrano's assumptions.

I'm going to investigate other solutions for now.