Provide parameterized templates for individual openswan connections & secrets.
Add newlines at the end of configuration files to avoid parser errors. Note that individual configuration files in /etc/ipsec.d/ are concatenated by the include statement.

We also noticed that puppet didn't seem to notice when the ipsec service couldn't reload it's configuration. This was due to a bug in the openswan initd script. We've changed our puppet recipe to patch the initd script after installation. Since we're patching a file, we thought it was sensible to fix the version of the openswan package.

Patch submitted to openswan - https://gsoc.xelerance.com/issues/1160