Easily add iptable rules from application puppet manifests.
Whenever defining an iptables rule from an application we must require the pre and post iptables files to be created otherwise rules are created on a blank server incorrectly. I couldn't find another way that worked to correctly ensure the order of loading the iptables rules.

Using an existing puppet module to handle iptable definition, checking and reloading[1]. Better creates the iptable rules using the name from puppet as a comment.

[1] http://bob.sh/puppet-iptables