Run Passenger and Rack processes as "application" user instead of "freerange" to bring them into line with application daemon processes.
I had hoped this would avoid the following warning/error that is appearing in the resque_worker & resque_scheduler daemon log output on a freshly puppet-provisioning VM :-

    Rails Error: Unable to access log file. Please ensure that /var/apps/hashblue/releases/20110121121403/log/staging.log exists and is chmod 0666. The log level has been raised to WARN and the output directed to STDERR until the problem is fixed.

It seems that although this commit changes the owner of the Passenger and Rack processes, the owner of staging.log and exceptional.log seems to still be "freerange". I think this is still a worthwhile change, but further investigation is required.